# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Instantly Service",
    "description": "Cold email outreach service via Instantly.ai API. Manages campaigns, leads, accounts, analytics, and webhook events.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3011"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AddLeadsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "added": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "success",
          "added",
          "total"
        ]
      },
      "AddLeadsRequest": {
        "type": "object",
        "properties": {
          "leads": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "companyName": {
                  "type": "string"
                },
                "customVariables": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "email"
              ]
            }
          }
        },
        "required": [
          "leads"
        ]
      },
      "DeleteLeadsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "deleted": {
            "type": "number"
          }
        },
        "required": [
          "success",
          "deleted"
        ]
      },
      "DeleteLeadsRequest": {
        "type": "object",
        "properties": {
          "emails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "emails"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/campaigns/{campaignId}/leads": {
      "post": {
        "summary": "Add leads to a campaign",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLeadsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leads added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddLeadsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Failed to add leads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List campaign leads",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "skip",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Leads list"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete leads from a campaign",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "campaignId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Campaign ID — automatically injected by workflow-service on all DAG calls",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s) — comma-separated UUIDs, automatically injected by workflow-service on all DAG calls. Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug — automatically injected by workflow-service on all DAG calls"
            },
            "required": false,
            "description": "Workflow slug — automatically injected by workflow-service on all DAG calls",
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug — propagated through the full call chain for tracking"
            },
            "required": false,
            "description": "Feature slug — propagated through the full call chain for tracking",
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteLeadsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leads deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteLeadsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing emails",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}